DORC scATAC-seq score summarization for single cells

Function to compute single cell DORC scores for each gene identified as a DORC (see runGenePeakcorr for determining DORCs through peak-gene correlations)

Usage

getDORCScores(
  ATAC.se,
  dorcTab,
  normalizeATACmat = TRUE,
  geneList = NULL,
  nCores = 4
)

Arguments

ATAC.se

SummarizedExperiment object of the scATAC-seq reads in peak counts.

dorcTab

data.frame object containing significant peak-gene pairs using which DORC scores will be computed. Must be a filtered set returned from runGenePeakcorr. IMPORTANT: Make sure the exact same scATAC SE (peak set) was used when determining DORCs that is used here to get corresponding DORC peak counts

normalizeATACmat

boolean indicating whether or not to normalize the counts present in the ATAC.se object prior to summing peak counts per gene. Default is TRUE (i.e. assumes peak counts are raw).

geneList

character vector specifying a subset of genes to compute scores for. Useful if you already have a set list of DORCs and you only want to compute scores for those genes/peaks.

nCores

numeric indicating the number of cores to use if parallelizing tasks

Value

a Matrix object of scores pertaining to the sum of normalized scATAC peak counts per gene, as determined for significant gene-peak pairs when calling DORCs

Author

Vinay Kartha