Infer FigR TF-DORC associations

Function to run TF motif-to-gene associations using reference DORC peak-gene mappings and TF RNA expression levels

Usage

runFigRGRN(
  ATAC.se,
  dorcK = 30,
  dorcTab,
  n_bg = 50,
  genome,
  dorcMat,
  rnaMat,
  dorcGenes = NULL,
  nCores = 1
)

Arguments

ATAC.se

SummarizedExperiment object of peak x cell scATAC-seq data, the same as used to compute DORCs using runGenePeakcorr

dorcK

numeric specifying the number of dorc nearest-neighbors to pool peaks from for the motif enrichment per DORC. Default is 30, i.e. set to ~3 percent of total DORCs determined

dorcTab

data.frame object containing significant peak-gene pairs using which DORC scores will be computed. Must be a filtered set returned from runGenePeakcorr. IMPORTANT: Make sure the exact same scATAC SE peak set was used when determining DORCs that is used here to get corresponding DORC peak counts

n_bg

number of background peaks to use for

genome

character specifying a valid genome assembly to use for peak GC content estimation and background peak determination. Must be one of "hg19","hg38", or "mm10", and requires the corresponding genomes package e.g. BSgenome.Hsapiens.UCSC.hg19 for hg19

dorcMat

Matrix object of smoothed single-cell DORC accessibility scores

rnaMat

Matrix object of smoothed single-cell RNA expression values

dorcGenes

character vector specifying the subset of DORCs to test, if not running on everything

nCores

numeric specifying the number of cores to run DORCs in parallel. Default is 1, i.e. don't use parallel backend

Value

a data.frame with all TF-DORC motif enrichment and correlation associations, and the corresponding FigR regulation score for each association

Author

Vinay Kartha